+++ build.sbt
@@ -190,7 +190,7 @@ lazy val redis = monixConnector("redis", Dependencies.Redis)
 
 lazy val s3 = monixConnector("s3", Dependencies.S3)
 
-lazy val gcs = monixConnector("google-cloud-storage", Dependencies.GCS)
+lazy val gcs = monixConnector("gcs", Dependencies.GCS)
 
 def monixConnector(
   connectorName: String,
