+++ project/Dependencies.scala
@@ -11,12 +11,13 @@ object Dependencies {
     val AkkaStreams = "2.6.5"
     val Hadoop = "3.1.3"
     val GCS = "1.107.0"
-
+    val Cats_Effect = "2.1.3"
     //test
     val Scalatest = "3.1.2"
     val Scalacheck = "1.14.0"
     val Mockito = "1.14.3"
     val Cats = "2.0.0"
+    val GCNio = "0.121.2"
   }
 
   private def commonDependencies(hasIntegrationTest: Boolean = false): Seq[sbt.ModuleID] = {
@@ -88,13 +89,12 @@ object Dependencies {
 
   val Redis = RedisDependencies ++ CommonProjectDependencies ++ CommonTestDependencies.map(_ % Test)
 
-
   private val GcsDependencies = Seq(
     "org.typelevel"     %% "cats-core"            % DependencyVersions.Cats,
-    "com.google.cloud"   % "google-cloud-storage" % "1.107.0",
-    "com.google.cloud" % "google-cloud-nio" % "0.121.2",
-    "org.typelevel" %% "cats-effect" % "2.1.3",
-   "commons-io" % "commons-io" % "2.6" % Test
+    "com.google.cloud"   % "google-cloud-storage" % DependencyVersions.GCS,
+    "org.typelevel" %% "cats-effect" % DependencyVersions.Cats_Effect,
+    "com.google.cloud" % "google-cloud-nio" % DependencyVersions.GCNio % IntegrationTest,
+    "commons-io" % "commons-io" % "2.6" % Test
   )
 
   val GCS = GcsDependencies  ++ CommonProjectDependencies ++ CommonTestDependencies
